removing certain files andfolder to organize the workplace with the new pages created (Home and Login pages)
