--- my-bq-app/src/pages/BreakfastPage.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-import MediaCard from '../components/BreakfastMediaCard.js';
-
-
-const BreakfastPage = () => {
-    return (
-     <MediaCard></MediaCard>
-    )
-  }
-  
-  export default BreakfastPage;
