--- my-bq-app/src/pages/HomePage.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-import MediaCard from '../components/MediaCard.js';
-
-
-const HomePage = () => {
-    return (
-      <MediaCard></MediaCard>
-    )
-  }
-  
-  export default HomePage;
