--- my-bq-app/src/pages/LunchPage.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-import MediaCard from '../components/BreakfastMediaCard.js';
-
-
-const LunchPage = () => {
-    return (
-     <MediaCard></MediaCard>
-    )
-  }
-  
-  export default LunchPage;
