+++ my-bq-app/src/styles/App.css
@@ -5,11 +5,13 @@
   box-sizing: border-box;
 }
 body {
-  background-image: url("../assets/wallpaper3.jpg");
+  background: rgba(249, 178, 132, 0.1);
+  mix-blend-mode: hard-light;
   background-size: cover;
   background-repeat: no-repeat;
   width: 100vw;
   height: 100vh;
+  position:relative;
 }
